# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Illuminae API Documentation",
        "description": "",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://illuminate.io-devs.us/",
            "description": "Staging API Server"
        },
        {
            "url": "https://illuminate.io-devs.us/",
            "description": "Live API Server"
        },
        {
            "url": "http://backend_illuminate.test/",
            "description": "Local Server"
        },
        {
            "url": "http://192.168.0.143:8009/",
            "description": "Local Server Using IP"
        }
    ],
    "paths": {
...
        "/api/user/profile/getUser": {
            "get": {
                "tags": [
                    "user,profile,getUser"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/updatePassword": {
            "post": {
                "tags": [
                    "user,profile,updatePassword"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "updatePassword",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "query",
                        "description": "Old Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/uploadPhoto": {
            "post": {
                "tags": [
                    "user,profile,uploadPhoto"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "uploadPhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "description": "Photo",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/storeUsername": {
            "post": {
                "tags": [
                    "user,profile,storeUsername"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "storeUsername",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/checkUsernameAvailability": {
            "post": {
                "tags": [
                    "user,profile,checkUsernameAvailability"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "checkUsernameAvailability",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/storeGeneralData": {
            "post": {
                "tags": [
                    "user,profile,storeGeneralData"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "storeGeneralData",
                "parameters": [
                    {
                        "name": "age",
                        "in": "query",
                        "description": "Age",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "description": "Height",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "about_you",
                        "in": "query",
                        "description": "About you",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_no",
                        "in": "query",
                        "description": "Phone Number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "facebook_url",
                        "in": "query",
                        "description": "Facebook Url",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instagram_url",
                        "in": "query",
                        "description": "Instagram Url",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "twitter_url",
                        "in": "query",
                        "description": "Twitter Url",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/profile/storeLocation": {
            "post": {
                "tags": [
                    "user,profile,storeLocation"
                ],
                "summary": "user profile",
                "description": "",
                "operationId": "storeLocation",
                "parameters": [
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "latitude",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}